% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{change_next_workflow_step}
\alias{change_next_workflow_step}
\title{Alter the Next Step of a Running Workflow}
\usage{
change_next_workflow_step(next_step, sbatch_opts = NULL)
}
\arguments{
\item{next_step}{A scalar number coercible to integer instructing the
workflow system which step to run next}

\item{sbatch_opts}{an optional named list of sbatch parameters that would
override the ones specified by the \code{next_step} for the next iteration}
}
\value{
The \code{next_step} value (invisibly)
}
\description{
This function allows a running job to alter the workflow sequence by choosing
which step to run after the current one. (See
\code{vignette("altering-the-order-of-steps")})
}
\examples{
\dontrun{
 # Instruct the workflow to run the step 3 after this one
 change_next_workflow_step(3)

 # Instruct the workflow to run the previous step after this one
 change_next_workflow_step(get_current_workflow_step() - 1)
}

}
