% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{get_current_workflow_step}
\alias{get_current_workflow_step}
\title{Get the Number of the Currently Running Step of a Workflow}
\usage{
get_current_workflow_step()
}
\value{
The \code{current_step} value
}
\description{
Get the Number of the Currently Running Step of a Workflow
}
