% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{get_workflow_root}
\alias{get_workflow_root}
\title{Get the root directory of a workflow}
\usage{
get_workflow_root(wf_summary = NULL)
}
\arguments{
\item{wf_summary}{The workflow summary returned after it's creation or
obtained with \code{swf_load_workflow}. If not provided, the function assumes
that the workflow is running on an HPC and pulls the value using
environment variables.}
}
\value{
The path to the workflow root directory
}
\description{
This function will get the path to the root directory of a workflow. Either
a local workflow or during the execution of a workflow on an HPC
}
