% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_templates.R
\name{helper_use_setup_lines}
\alias{helper_use_setup_lines}
\title{Helper Function to Consistently Write Setup Lines}
\usage{
helper_use_setup_lines(instructions, setup_lines)
}
\arguments{
\item{instructions}{Vector of bash lines to be run by the workflow step}

\item{setup_lines}{Vector of bash lines to be run before the rest of the
step instructions.

@return The modified instructions}
}
\description{
Helper Function to Consistently Write Setup Lines
}
