% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_templates.R
\name{helper_write_instructions}
\alias{helper_write_instructions}
\title{Helper function to consistently write instructions to the instruction script}
\usage{
helper_write_instructions(instructions, instructions_script)
}
\arguments{
\item{instructions}{Vector of bash lines to be run by the workflow step}

\item{instructions_script}{Path to the instructions script}
}
\description{
Helper function to consistently write instructions to the instruction script
}
