% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simple_brew}
\alias{simple_brew}
\title{Replace a line in a file with new lines}
\usage{
simple_brew(file_in, placeholder_line, replacement_lines, file_out = NULL)
}
\arguments{
\item{file_in}{the file to read from}

\item{placeholder_line}{the line to replace}

\item{replacement_lines}{a character vector of lines to insert into the file}

\item{file_out}{path to the new file}
}
\value{
path to the output file (invisibly)
}
\description{
This function is used to complete the templates provided by the package with
the user provided informations
}
