% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurmworkflow-package.R
\docType{package}
\name{slurmworkflow-package}
\alias{slurmworkflow}
\alias{slurmworkflow-package}
\title{slurmworkflow: Build workflows for the Slurm workload manager}
\description{
slurmworkflow makes it possible to locally pre-define a set of jobs to be run on a slurm equipped HPC. A workflow is a set of steps to be run on the HPC (sequentially by default). slurmworkflow provides tools to alter the execution order, allowing conditional execution of the steps and loop like behavior.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://epimodel.github.io/slurmworkflow/}
}

}
\author{
\strong{Maintainer}: Adrien Le Guillou \email{contact@aleguillou.org} (\href{https://orcid.org/0000-0002-4791-418X}{ORCID})

}
\keyword{internal}
