% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_templates.R
\name{step_tmpl_bash_script}
\alias{step_tmpl_bash_script}
\title{Step Template to Run a Bash Script}
\usage{
step_tmpl_bash_script(bash_script)
}
\arguments{
\item{bash_script}{Path to the script to be run by the workflow step}
}
\value{
a template function to be used by \code{add_workflow_step}
}
\description{
Step Template to Run a Bash Script
}
\section{Step Template}{

Step Templates are helper functions to be used within \code{add_workflow_step}.
Some basic ones are provided by the \code{slurmworkflow} package. They instruct
the workflow to run either a bash script, a set of bash lines given as a
character vector or an R script.
Additional Step Templates can be created to simplify specific tasks, see the
\code{vignette("making-a-custom-step-template")} for details.
}

