% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_templates.R
\name{step_tmpl_map_script}
\alias{step_tmpl_map_script}
\title{Step template to run an R script with a set of arguments}
\usage{
step_tmpl_map_script(r_script, ..., MoreArgs = NULL, setup_script = NULL)
}
\arguments{
\item{r_script}{The R script to be run by the workflow step}

\item{...}{arguments to vectorize over (vectors or lists of strictly
positive length, or all of zero length).  See also ‘Details’.}

\item{MoreArgs}{a \emph{list} of arguments to the function call. The \code{names}
attribute of \code{args} gives the argument names.}

\item{setup_script}{(optional) a bash script to be run first.
This can be used to load the required modules (like R, python, etc).}
}
\value{
a template function to be used by \code{add_workflow_step}
}
\description{
Step template to run an R script with a set of arguments
}
\section{Step Template}{

Step Templates are helper functions to be used within \code{add_workflow_step}.
Some basic ones are provided by the \code{slurmworkflow} package. They instruct
the workflow to run either a bash script, a set of bash lines given as a
character vector or an R script.
Additional Step Templates can be created to simplify specific tasks, see the
\code{vignette("making-a-custom-step-template")} for details.
}

