% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_templates.R
\name{step_tmpl_renv_restore}
\alias{step_tmpl_renv_restore}
\title{Step template to update a project \code{renv}}
\usage{
step_tmpl_renv_restore(setup_lines = NULL)
}
\arguments{
\item{setup_lines}{(optional) a vector of bash lines to be run first.
This can be used to load the required modules (like R, python, etc).}
}
\value{
a template function to be used by \code{add_workflow_step}
}
\description{
This template makes the step run \verb{git pull} and \code{renv::restore()}. This could
help ensure that the project is up to date when running the rest of the
workflow.
}
\section{Step Template}{

Step Templates are helper functions to be used within \code{add_workflow_step}.
Some basic ones are provided by the \code{slurmworkflow} package. They instruct
the workflow to run either a bash script, a set of bash lines given as a
character vector or an R script.
Additional Step Templates can be created to simplify specific tasks, see the
\code{vignette("making-a-custom-step-template")} for details.
}

