% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{get_current_workflow_step}
\alias{get_current_workflow_step}
\title{Get the number of the currently running step of a workflow}
\usage{
get_current_workflow_step()
}
\value{
The \code{current_step} value
}
\description{
Get the number of the currently running step of a workflow
}
