% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{load_workflow}
\alias{load_workflow}
\title{Load a Workflow Summary From a Workflow Folder or the Environment}
\usage{
load_workflow(wf_root = NULL)
}
\arguments{
\item{wf_root}{Path to a workflow directory. If not provided, the function
assumes that the workflow is running on an HPC and pulls the value using
environment variables.}
}
\value{
The workflow summary
}
\description{
Load a Workflow Summary From a Workflow Folder or the Environment
}
