% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_templates.R
\name{step_tmpl_map_script}
\alias{step_tmpl_map_script}
\title{Step Template to Run an R Script With a Set of Arguments}
\usage{
step_tmpl_map_script(r_script, ..., MoreArgs = NULL, setup_lines = NULL)
}
\arguments{
\item{r_script}{The R script to be run by the workflow step}

\item{...}{arguments to vectorize over (vectors or lists of strictly
positive length, or all of zero length).  See also ‘Details’.}

\item{MoreArgs}{a \code{list} of arguments to the function call. The \emph{names}
attribute of \code{args} gives the argument names.}

\item{setup_lines}{(optional) a vector of bash lines to be run first.
This can be used to load the required modules (like R, python, etc).}
}
\value{
a template function to be used by \code{add_workflow_step}
}
\description{
Step Template to Run an R Script With a Set of Arguments
}
\section{Step Template}{

Step Templates are helper functions to be used within \code{add_workflow_step}.
Some basic ones are provided by the \code{slurmworkflow} package. They instruct
the workflow to run either a bash script, a set of bash lines given as a
character vector or an R script.
Additional Step Templates can be created to simplify specific tasks, see the
\code{vignette("making-a-custom-step-template")} for details.
}

